api_services/requirements.txt
# Data Sources REST API - Requirements

# Core
python-dotenv
pydantic
typing_extensions

# API
fastapi
uvicorn

# LangChain & Graph
langchain
langchain-core
langchain-community
langchain-openai
langchain-mcp-adapters
langgraph
langchain-tavily
langchain-groq
langgraph-supervisor
tavily-python
langchain-qdrant
langgraph-checkpoint-postgres
langgraph-checkpoint-sqlite
aiosqlite==0.21.0
# Vector DB
faiss-cpu

# PDF & Images
pillow
pypdf
pymuPDF

# Analysis & Table Extraction
beautifulsoup4
unstructured-client
openpyxl
pandas
matplotlib
reportlab
seaborn

# OpenAI client
openai

# Utilities
requests
python-multipart
qdrant-client
llama_parse

# Google APIs
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib

# MCP
fastmcp

# UI

psycopg2-binary

# Pin websockets to avoid deprecation warning
websockets